+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* remove Freemius from the autoloader
 
 #### 2.3.1 / 2021-06-14
 * utilize new `class Ignore` in Git Updater