+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 2.4.0 / 2021-07-05
 * remove Freemius from the autoloader
 * utilize new `class Ignore` from Git Updater
 * uses new `class Fragen\Git_Updater\Shim` for PHP 5.6 compatibility, will remove when WP core changes minimum requirement